globalpoiseidon-eclipse-formatter.xml hinzugefügt. Bitte in eclipse laden als Formatter und vor jedem Commit auf das Projekt anwenden. Max linelength wurde auf 140 erhöht

+++ src/de/globalposeidon/Qualitaet/tests/BuildingTest.java
@@ -1,7 +1,5 @@
 package de.globalposeidon.Qualitaet.tests;
 
-import java.util.ArrayList;
-
 import junit.framework.TestCase;
 
 import org.junit.After;
@@ -17,179 +15,178 @@ import de.globalposeidon.Qualitaet.model.Renter;
  * This class contains the tests for the building class.
  *
  * @author Krischan Stapelfeldt
- *
  */
 public class BuildingTest extends TestCase {
 
-	private DataContainer dataContainer;
-	private Renter renter;
-	private Entrance entrance;
-//	private ArrayList<Entrance> entrances;
-	private Building building;
+   private DataContainer dataContainer;
+   private Renter renter;
+   private Entrance entrance;
+   // private ArrayList<Entrance> entrances;
+   private Building building;
 
-	@Override
-	@Before
-	/**
-	 * Ramps the necessary objects up (dataContainer, renter, building,
-	 * entrance).
-	 * @throws Throws an exception, when catching fire.
-	 */
-	public final void setUp() throws Exception {
-		dataContainer = new DataContainer();
-		renter = new Renter();
-		building = new Building(dataContainer);
-		entrance = new Entrance(building);
-//		entrances = new ArrayList<Entrance>();
-//		entrances.add(initialEntrance);
-	}
+   @Override
+   @Before
+   /**
+    * Ramps the necessary objects up (dataContainer, renter, building,
+    * entrance).
+    * @throws Throws an exception, when catching fire.
+    */
+   public final void setUp() throws Exception {
+      dataContainer = new DataContainer();
+      renter = new Renter();
+      building = new Building(dataContainer);
+      entrance = new Entrance(building);
+      // entrances = new ArrayList<Entrance>();
+      // entrances.add(initialEntrance);
+   }
 
-	@Override
-	@After
-	public void tearDown() throws Exception {
-	}
+   @Override
+   @After
+   public void tearDown() throws Exception {
+   }
 
-	/**
-	 * create a new building with specific data. insert into database.
-	 */
-	@Test
-	public final void testPositiveCreateBuilding() {
+   /**
+    * create a new building with specific data. insert into database.
+    */
+   @Test
+   public final void testPositiveCreateBuilding() {
 
-		assertNotNull(building);
-	}
+      assertNotNull(building);
+   }
 
-	/**
-	 * bla.
-	 */
-	@Test
-	public final void testGetID() {
+   /**
+    * bla.
+    */
+   @Test
+   public final void testGetID() {
 
-		assertNotNull((Integer) building.getID());
-	}
+      assertNotNull(building.getID());
+   }
 
-	/**
-	 * set renter to building.
-	 */
-	@Test
-	public final void testPositiveSetRenter() {
+   /**
+    * set renter to building.
+    */
+   @Test
+   public final void testPositiveSetRenter() {
 
-		building.setRenter(renter);
-		assertEquals("", renter, building.getRenter());
+      building.setRenter(renter);
+      assertEquals("", renter, building.getRenter());
 
-	}
+   }
 
-	/**
-	 * get renter from selected building. load from database.
-	 */
-	@Test
-	public final void testPositiveGetRenter() {
+   /**
+    * get renter from selected building. load from database.
+    */
+   @Test
+   public final void testPositiveGetRenter() {
 
-		building.setRenter(renter);
-		assertEquals("", renter, building.getRenter());
+      building.setRenter(renter);
+      assertEquals("", renter, building.getRenter());
 
-	}
+   }
 
-	/**
-	 * replace renter with a new renter. insert into database.
-	 */
-	@Test
-	public final void testPositiveReplaceRenter() {
+   /**
+    * replace renter with a new renter. insert into database.
+    */
+   @Test
+   public final void testPositiveReplaceRenter() {
 
-		final Renter newRenter = new Renter();
-		building.setRenter(newRenter);
-		assertEquals("", newRenter, building.getRenter());
+      final Renter newRenter = new Renter();
+      building.setRenter(newRenter);
+      assertEquals("", newRenter, building.getRenter());
 
-	}
+   }
 
-	/**
-	 * add entrance to selected building. insert into database.
-	 */
-	@Test
-	public final void testPositiveAddEntrance() {
+   /**
+    * add entrance to selected building. insert into database.
+    */
+   @Test
+   public final void testPositiveAddEntrance() {
 
-		building.addEntrance(entrance);
-		assertEquals("", entrance, building.getEntrance(0));
+      building.addEntrance(entrance);
+      assertEquals("", entrance, building.getEntrance(0));
 
-	}
+   }
 
-	/**
-	 * get entrance from selected ArrayList. load from database.
-	 */
-	@Test
-	public final void testPositiveGetEntrance() {
+   /**
+    * get entrance from selected ArrayList. load from database.
+    */
+   @Test
+   public final void testPositiveGetEntrance() {
 
-		building.addEntrance(entrance);
-		assertEquals("", entrance, building.getEntrance(0));
+      building.addEntrance(entrance);
+      assertEquals("", entrance, building.getEntrance(0));
 
-	}
+   }
 
-	/**
-	 * get entrance from selected ArrayList. load from database.
-	 */
-	@Test
-	public final void testNegativeGetEntrance() {
+   /**
+    * get entrance from selected ArrayList. load from database.
+    */
+   @Test
+   public final void testNegativeGetEntrance() {
 
-		assertEquals("", null, building.getEntrance(0));
+      assertEquals("", null, building.getEntrance(0));
 
-	}
+   }
 
-	/**
-	 * remove entrance from selected ArrayList. delete from database.
-	 */
-	@Test
-	public final void testPositiveRemoveEntrance() {
+   /**
+    * remove entrance from selected ArrayList. delete from database.
+    */
+   @Test
+   public final void testPositiveRemoveEntrance() {
 
-		building.addEntrance(entrance);
-		assertEquals("", entrance, building.removeEntrance(0));
+      building.addEntrance(entrance);
+      assertEquals("", entrance, building.removeEntrance(0));
 
-	}
+   }
 
-	/**
-	 * remove entrance from selected ArrayList. delete from database.
-	 */
-	@Test
-	public final void testNegativeRemoveEntrance() {
+   /**
+    * remove entrance from selected ArrayList. delete from database.
+    */
+   @Test
+   public final void testNegativeRemoveEntrance() {
 
-		assertEquals("", null, building.removeEntrance(0));
+      assertEquals("", null, building.removeEntrance(0));
 
-	}
+   }
 
-	/**
-	 * get empty buildings from selected ArrayList. load from database.
-	 */
-	@Test
-	public final void testPositiveBuildingIsEmpty() {
+   /**
+    * get empty buildings from selected ArrayList. load from database.
+    */
+   @Test
+   public final void testPositiveBuildingIsEmpty() {
 
-		assertEquals("", true, building.buildingIsEmpty());
-	}
+      assertEquals("", true, building.buildingIsEmpty());
+   }
 
-	/**
-	 * bla.
-	 */
-	@Test
-	public final void testNegativeBuildingIsEmpty() {
+   /**
+    * bla.
+    */
+   @Test
+   public final void testNegativeBuildingIsEmpty() {
 
-		building.addEntrance(entrance);
-		assertEquals("", false, building.buildingIsEmpty());
-	}
+      building.addEntrance(entrance);
+      assertEquals("", false, building.buildingIsEmpty());
+   }
 
-	/**
-	 * bla.
-	 */
-	@Test
-	public final void testEntranceCount() {
+   /**
+    * bla.
+    */
+   @Test
+   public final void testEntranceCount() {
 
-		assertEquals("", 0, building.entranceCount());
-	}
+      assertEquals("", 0, building.entranceCount());
+   }
 
-	/**
-	 * bla.
-	 */
-	@Test
-	public final void testToString() {
+   /**
+    * bla.
+    */
+   @Test
+   public final void testToString() {
 
-		assertNotNull(building.toString());
-	}
-// <----- TREENODE TESTS ----->
+      assertNotNull(building.toString());
+   }
+   // <----- TREENODE TESTS ----->
 
-// TODO: Write this tests?
+   // TODO: Write this tests?
 }
